Mudança de tipo de variavel
Score_do_paciente, troquei de float para int

+++ Database/BancodeDadosAndrey.sql
@@ -37,7 +37,7 @@ create table historico_de_consulta (
   CPF_Paciente varchar(11),
   CPF_Responsavel varchar(11),
   id_medico int,
-  score_do_paciente float,
+  score_do_paciente int,
   sintomas varchar(250),
   FOREIGN KEY (CPF_Paciente) REFERENCES Paciente(CPF_Paciente),
   FOREIGN KEY (CPF_Responsavel) REFERENCES Responsavel(CPF_Responsavel),
